Fix test failure for builds in source tree
The fs::snapsot_t class creates a finger print of a folder. We used the source tree for testing, but if the source tree hosts the build directory, then the fingerprint will (likely) change, since we run simultaneous build jobs.

It now uses the Frameworks folder in the source tree, as it’s unlikely that someone would place their build directory in this location.