Don’t mark empty snippet placeholders ‘unanchored’
Ranges can be unanchored which is when they are made (mainly) by a unit selection, for example “select to enclosing braces”. What it means is that if we later use shift + movement actions (to “move” the selection end point) we will always extend it (as we don’t know which of the two end points should be considered the anchor).

Though the unanchored flag has a double role, for snippet placeholders we also set it, when we select the placeholder field, and when user press tab, we ignore tab expansions if there is an unanchored selection. So this disables recursive snippet expansion if the user immediately press tab when reaching a snippet placeholder.

Turns out that when we reach the last field of a snippet, some users want a new snippet expansion.

This fix is not ideal, but by far the easiest, and it achieves the goal.

This closes issue #182.