Evaluate embedded ruby in markdown files
This is when generating HTML for the release notes, manual, and, where we actually need this, list of contributions.

Long-term we should probably switch to template tags as we are effectively adding code to the (otherwise declarative) build graph, which means we don’t have any way to tell if the generated HTML is up-to-date or not (as that would require analyzing the embedded ruby code).