Fix potential crash when using “Edit Command…” button
The block used the ‘uuid’ variable, this variable was however a (const) reference, so it wasn’t safe to use within the block.

While fixing the bug (by introducing a local non-reference variable for the bundle item) we have also changed the code slightly, so that “Edit Command…” only shows when a bundle item is found (this is mainly in preparation for running shell commands that doesn’t have a bundle item, e.g. from a future “Filter Through Command” dialog).