Workaround for flamboyant progress bar
Starting with Lion (I think) the progress bar at the Bundles preferences pane won’t always hide after having installed a bundle.

The progress bar is set to not show when stopped, but for some reason this property does not seem to always be respected, but if we manually set the progress to zero before setting the property (bound to ‘animate’) to NO, then it does properly hide.