Offer double-click install of plug-ins
If the plug-in doesn’t set TMPlugInAPIVersion to 2 (in its Info.plist) a dialog will show, telling the user that the plug-in is not compatible with the current version of TextMate (which is most likely correct).

You can hold down option (⌥) to open the plug-in as a regular folder.

Addresses issue #392.