Default file type is Objective-C++ for this project
Previously it was C++ — not sure if a setting like this belongs in the project or should be user-specific. Presently though the .tm_properties system lack a way to include user specific files, and while one can put the settings in ~/.tm_properties and make them target a (project) sub-folder, they would not override the settings set in the project.