Fix resize constraints for file browser / HTML output
These were set to ‘NSLayoutPriorityRequired-1’ which is higher than the window’s size constraints, meaning that resizing the window would not allow shrinking the file browser (or HTML output) views, furthermore, by making the views larger (when dragging the divider) the growth would not be constrained by the current window size, so it was possible to make the window grow with it.

The priority is now the proper NSLayoutPriorityDragThatCannotResizeWindow.