Don’t ask to save untitled documents during quit
This is only if there are no other documents with unsaved changes and the user has ‘open documents from last session’ enabled in preferences.

The rationale behind only offering it for untitled documents is that we can be sure that these will not be modified by another program while TextMate is not running. It also seems to be less valuable to have TextMate keep unsaved changes to a file that exist on disk, in fact, it’s a pretty bad behavior IMHO (and dangerous if other programs may work on the file).