add "Open in Finder" context menu when no items are selected
When you control-click the empty part of the file browser and select "Open in Finder" a Finder window will open showing the currently selected folder.

You can go about this other ways, but I find that this feature complements the already existing "New Folder" menu item in the no-items-selected context menu of the file browser.

This patch is public domain.